Backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cloudinary==1.24.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.1
whitenoise==5.2.0
psycopg2-binary>=2.8
django-import-export==2.4.0